Connect edit item to api endpoint
