+++ roomies-client/src/app/app.module.ts
@@ -17,7 +17,6 @@ import { JoinHouseComponent } from './join-house/join-house.component';
 import { CreateHouseComponent } from './create-house/create-house.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
 import { AddChoreTaskComponent } from './add-chore-task/add-chore-task.component';
-import { EditGroceryItemComponent } from './edit-grocery-item/edit-grocery-item.component';
 import { EditChoreTaskComponent } from './edit-chore-task/edit-chore-task.component';
 
 @NgModule({
@@ -32,7 +31,6 @@ import { EditChoreTaskComponent } from './edit-chore-task/edit-chore-task.compon
     CreateHouseComponent,
     DashboardComponent,
     AddChoreTaskComponent,
-    EditGroceryItemComponent,
     EditChoreTaskComponent
   ],
   imports: [
