--- roomies-client/src/app/edit-grocery-item/edit-grocery-item.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { EditGroceryItemComponent } from './edit-grocery-item.component';
-
-describe('EditGroceryItemComponent', () => {
-  let component: EditGroceryItemComponent;
-  let fixture: ComponentFixture<EditGroceryItemComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ EditGroceryItemComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(EditGroceryItemComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
